feat: add E2E test to live check the lib works
This is intended to be used with ts-node in development mode. It imports
the library and listens for a reaction on a specific message to add a
specific bot.

This code should be used to assert that the bot works. Provide the
token, the message ID and the role ID using environment variables, such
as

* BOT_TOKEN: for the login token
* ROLE: for the snowflake of the Role I should add
* MESSAGE: for the snowflake of the Message I should listen