feat: support parsing custom emojis by ID or name
This commit will also fix a behaviour that I took for granted. Of course
when a custom emoji is used, it is not possible to get the name, because
emoji.id is always set.

This commit replaces the config parsing logic so that it tries to find
a role by ID if present. If not present, or if present but not
configured, it tries to find a role by name, as usual.