refactor!: do not use an export default
This breaking change will force all the existing imports to change, but
it will be for good. This removes the default export, so now
ReactionRole is exported with its own name. This is also safer to
prevent using a different variable name and have unexpected results.

For CommonJS projects, this means probably using .default is not
required anymore, the following should be used instead:

    const { ReactionRole } = require("discordjs-reaction-role")

For TypeScript and ES Module projects, it means wrapping the import in
braces (and making sure that the proper name is being used):

    import { ReactionRole } from "discordjs-reaction-role"