refactor: allow extracting role from partials
It seems that partials already have all the required information to
handle the role extraction, since they don't need to be fetched given
that they already provide the message ID, the reaction emoji and the
user ID.

Therefore, in this commit I remove the fetch() calls made to the API to
always fetch the partial reactions, which should make it more performant
and also allows to remove the case where the message cannot be fetched
due to it being removed.

Hopefully, since the .deleted message was deprecated, this also means
that the bot is more futureproof from now on and let's see if this means
that it is both compatible with Discord.js 13 and Discord.js 14 as it is
right now.