feat: add debug mode
Running the bot with REACTION_DEBUG environment variable set will print
some information to stdout whenever a reaction is added or removed, so
that it is possible to check whether things are working or not.

While this is a development tool for me to help me when testing the
library, I am going to leave this in the pack for the time being to see
if it makes integrations easier.

Note that this is strictly opt-in and if the environment variable is not
set, nothing will happen.